Update build directory, add serve make target

+++ Makefile
@@ -4,5 +4,9 @@ build: deps
 deps:
 	npm install
 
+serve:
+	python -m http.server --directory _build 7777
+
 clean:
-	rm -rf node_modules
+	[[ -e "node_modules" ]] && rm -rf node_modules
+	[[ -e "_build" ]] && rm -rf _build
